splitting large files and organizing into folders
