+++ trulens_eval/trulens_eval/provider/hugs.py
@@ -1,0 +1,213 @@
+
+from typing import Dict
+from trulens_eval.trulens_eval.provider import Provider
+from trulens_eval.trulens_eval.provider.endpoint import Endpoint, HuggingfaceEndpoint
+
+
+# Cannot put these inside Huggingface since it interferes with pydantic.BaseModel.
+
+HUGS_SENTIMENT_API_URL = "https://api-inference.huggingface.co/models/cardiffnlp/twitter-roberta-base-sentiment"
+HUGS_TOXIC_API_URL = "https://api-inference.huggingface.co/models/martin-ha/toxic-comment-model"
+HUGS_CHAT_API_URL = "https://api-inference.huggingface.co/models/facebook/blenderbot-3B"
+HUGS_LANGUAGE_API_URL = "https://api-inference.huggingface.co/models/papluca/xlm-roberta-base-language-detection"
+HUGS_NLI_API_URL = "https://api-inference.huggingface.co/models/ynie/roberta-large-snli_mnli_fever_anli_R1_R2_R3-nli"
+HUGS_DOCNLI_API_URL = "https://api-inference.huggingface.co/models/MoritzLaurer/DeBERTa-v3-base-mnli-fever-docnli-ling-2c"
+
+
+class Huggingface(Provider):
+
+    endpoint: Endpoint
+
+    def __init__(self, endpoint=None, **kwargs):
+        # NOTE(piotrm): pydantic adds endpoint to the signature of this
+        # constructor if we don't include it explicitly, even though we set it
+        # down below. Adding it as None here as a temporary hack.
+        """
+        A set of Huggingface Feedback Functions.
+
+        All args/kwargs passed to HuggingfaceEndpoint constructor.
+        """
+
+        self_kwargs = dict()
+        self_kwargs['endpoint'] = HuggingfaceEndpoint(**kwargs)
+
+        super().__init__(
+            **self_kwargs
+        )  # need to include pydantic.BaseModel.__init__
+
+    def language_match(self, text1: str, text2: str) -> float:
+        """
+        Uses Huggingface's papluca/xlm-roberta-base-language-detection model. A
+        function that uses language detection on `text1` and `text2` and
+        calculates the probit difference on the language detected on text1. The
+        function is: `1.0 - (|probit_language_text1(text1) -
+        probit_language_text1(text2))`
+        
+        Parameters:
+        
+            text1 (str): Text to evaluate.
+
+            text2 (str): Comparative text to evaluate.
+
+        Returns:
+
+            float: A value between 0 and 1. 0 being "different languages" and 1
+            being "same languages".
+        """
+
+        def get_scores(text):
+            payload = {"inputs": text}
+            hf_response = self.endpoint.post(
+                url=HUGS_LANGUAGE_API_URL, payload=payload, timeout=30
+            )
+            return {r['label']: r['score'] for r in hf_response}
+
+        max_length = 500
+        scores1: AsyncResult[Dict] = TP().promise(
+            get_scores, text=text1[:max_length]
+        )
+        scores2: AsyncResult[Dict] = TP().promise(
+            get_scores, text=text2[:max_length]
+        )
+
+        scores1: Dict = scores1.get()
+        scores2: Dict = scores2.get()
+
+        langs = list(scores1.keys())
+        prob1 = np.array([scores1[k] for k in langs])
+        prob2 = np.array([scores2[k] for k in langs])
+        diff = prob1 - prob2
+
+        l1 = 1.0 - (np.linalg.norm(diff, ord=1)) / 2.0
+
+        return l1, dict(text1_scores=scores1, text2_scores=scores2)
+
+    def positive_sentiment(self, text: str) -> float:
+        """
+        Uses Huggingface's cardiffnlp/twitter-roberta-base-sentiment model. A
+        function that uses a sentiment classifier on `text`.
+        
+        Parameters:
+            text (str): Text to evaluate.
+
+        Returns:
+            float: A value between 0 and 1. 0 being "negative sentiment" and 1
+            being "positive sentiment".
+        """
+        max_length = 500
+        truncated_text = text[:max_length]
+        payload = {"inputs": truncated_text}
+
+        hf_response = self.endpoint.post(
+            url=HUGS_SENTIMENT_API_URL, payload=payload
+        )
+
+        for label in hf_response:
+            if label['label'] == 'LABEL_2':
+                return label['score']
+
+    def not_toxic(self, text: str) -> float:
+        """
+        Uses Huggingface's martin-ha/toxic-comment-model model. A function that
+        uses a toxic comment classifier on `text`.
+        
+        Parameters:
+            text (str): Text to evaluate.
+
+        Returns:
+            float: A value between 0 and 1. 0 being "toxic" and 1 being "not
+            toxic".
+        """
+        max_length = 500
+        truncated_text = text[:max_length]
+        payload = {"inputs": truncated_text}
+        hf_response = self.endpoint.post(
+            url=HUGS_TOXIC_API_URL, payload=payload
+        )
+
+        for label in hf_response:
+            if label['label'] == 'toxic':
+                return label['score']
+
+    def _summarized_groundedness(self, premise: str, hypothesis: str) -> float:
+        """ A groundedness measure best used for summarized premise against simple hypothesis.
+        This Huggingface implementation uses NLI.
+
+        Args:
+            premise (str): NLI Premise
+            hypothesis (str): NLI Hypothesis
+
+        Returns:
+            float: NLI Entailment
+        """
+        if not '.' == premise[len(premise) - 1]:
+            premise = premise + '.'
+        nli_string = premise + ' ' + hypothesis
+        payload = {"inputs": nli_string}
+        hf_response = self.endpoint.post(url=HUGS_NLI_API_URL, payload=payload)
+
+        for label in hf_response:
+            if label['label'] == 'entailment':
+                return label['score']
+
+    def _doc_groundedness(self, premise, hypothesis):
+        """ A groundedness measure for full document premise against hypothesis.
+        This Huggingface implementation uses DocNLI. The Hypoethsis still only works on single small hypothesis.
+
+        Args:
+            premise (str): NLI Premise
+            hypothesis (str): NLI Hypothesis
+
+        Returns:
+            float: NLI Entailment
+        """
+        nli_string = premise + ' [SEP] ' + hypothesis
+        payload = {"inputs": nli_string}
+        hf_response = self.endpoint.post(
+            url=HUGS_DOCNLI_API_URL, payload=payload
+        )
+
+        for label in hf_response:
+            if label['label'] == 'entailment':
+                return label['score']
+
+
+class Cohere(Provider):
+    model_engine: str = "large"
+
+    def __init__(self, model_engine='large', endpoint=None, **kwargs):
+        # NOTE(piotrm): pydantic adds endpoint to the signature of this
+        # constructor if we don't include it explicitly, even though we set it
+        # down below. Adding it as None here as a temporary hack.
+
+        kwargs['endpoint'] = Endpoint(name="cohere")
+        kwargs['model_engine'] = model_engine
+
+        super().__init__(
+            **kwargs
+        )  # need to include pydantic.BaseModel.__init__
+
+    def sentiment(
+        self,
+        text,
+    ):
+        return int(
+            Cohere().endpoint.run_me(
+                lambda: get_cohere_agent().classify(
+                    model=self.model_engine,
+                    inputs=[text],
+                    examples=feedback_prompts.COHERE_SENTIMENT_EXAMPLES
+                )[0].prediction
+            )
+        )
+
+    def not_disinformation(self, text):
+        return int(
+            Cohere().endpoint.run_me(
+                lambda: get_cohere_agent().classify(
+                    model=self.model_engine,
+                    inputs=[text],
+                    examples=feedback_prompts.COHERE_NOT_DISINFORMATION_EXAMPLES
+                )[0].prediction
+            )
+        )
